homebasket_backend/requirements.txt
Django>=4.2
djangorestframework>=3.14
mysqlclient>=2.2
django-cors-headers>=4.3
python-decouple>=3.8
django-extensions>=3.2